multi channel mel -win320 作为输入 网络最后的激活函数由softmax换为sigmoid
改frame_level_label，用卷积方式实现

+++ dataloader.py
@@ -27,7 +27,7 @@ class Dataset(Dataset):
 
         sample=(
             Variable(torch.FloatTensor(wav.astype('float32'))),
-            Variable(torch.LongTensor(label.astype('int64'))),
+            Variable(torch.LongTensor(label)),
             alpha_pow
         )
         return sample
